+++ src/common/logging.cpp
@@ -26,11 +26,13 @@ std::shared_ptr<spdlog::logger> createStderrLogger(const std::string& name,
                                                    const std::string& pattern,
                                                    const std::vector<std::string>& files,
                                                    bool quiet) {
-  std::vector<spdlog::sink_ptr> sinks;
+  auto logger = spdlog::get(name);
+  if(!logger) {
+    std::vector<spdlog::sink_ptr> sinks;
 
-  auto stderr_sink = spdlog::sinks::stderr_sink_mt::instance();
-  if(!quiet)
-    sinks.push_back(stderr_sink);
+    auto stderr_sink = spdlog::sinks::stderr_sink_mt::instance();
+    if(!quiet)
+      sinks.push_back(stderr_sink);
 
   // @TODO: think how to solve this better than using OMPI_COMM_WORLD_RANK env variable
   // only create output files if we are the main process or if MPI rank is not defined
@@ -42,10 +44,11 @@ std::shared_ptr<spdlog::logger> createStderrLogger(const std::string& name,
     }
   }
 
-  auto logger = std::make_shared<spdlog::logger>(name, begin(sinks), end(sinks));
+    logger = std::make_shared<spdlog::logger>(name, begin(sinks), end(sinks));
 
-  spdlog::register_logger(logger);
-  logger->set_pattern(pattern);
+    spdlog::register_logger(logger);
+    logger->set_pattern(pattern);
+  }
   return logger;
 }
 
@@ -72,6 +75,7 @@ bool setLoggingLevel(spdlog::logger& logger, std::string const level) {
 }
 
 static void setErrorHandlers();
+
 void createLoggers(const marian::Config* config) {
   std::vector<std::string> generalLogs;
   std::vector<std::string> validLogs;